Make project persist including networks + when changing project name in url keep the current route hash
