+++ src/components/UploadButton.tsx
@@ -2,33 +2,18 @@ import { UploadIcon } from "lucide-react";
 import { Button } from "./ui/button";
 import { useRef } from "react";
 import { store } from "@/lib/store";
-import { store as componentGraphStore } from "@/lib/graph-store/component";
-import { store as statementGraphStore } from "@/lib/graph-store/statement";
 import { useToast } from "@/hooks/use-toast";
 import { csvParse } from "d3-dsv";
 import { read as readXLSX, utils as utilsXLSX } from "xlsx";
 import { Statements, statementsSchema } from "@/lib/schema";
 import { load } from "@/lib/io";
+import { json2project } from "@/lib/project2json";
 import { ZodError } from "zod";
 
 async function processJSONFile(file: File) {
   const content = await file.text();
-  const state = JSON.parse(content);
-  // TODO validate state using zod
-  componentGraphStore.setState({
-    nodes: state.graph.component.nodes,
-    edges: state.graph.component.edges,
-  });
-  statementGraphStore.setState({
-    nodes: state.graph.statement.nodes,
-    edges: state.graph.statement.edges,
-  });
-  store.setState({
-    projectName: projectNameFromFile(file),
-    statements: state.statements,
-    connections: state.connections,
-    conflicts: state.conflicts,
-  });
+  const projectName = projectNameFromFile(file);
+  json2project(content, projectName);
 }
 
 /**
